+++ setup.cfg
@@ -23,7 +23,7 @@ markers =
     unit: marks unit tests (deselect with '-m "not unit"')
     integration: marks integration tests (deselect with '-m "not integration"')
     system: marks integration tests (deselect with '-m "not system"')
-    acceptance: marks integration tests (deselect with '-m "not acceptance"')
+    acceptance: marks acceptance tests (deselect with '-m "not acceptance"')
     docs: mark tests related to documentation (deselect with '-m "not docs"')
