+++ tests/acceptance/test_acceptance_speech_commands.py
@@ -19,14 +19,15 @@ import os
 import shutil
 import tarfile
 import unittest
+from unittest import TestCase
 
+import pytest
 from ruamel.yaml import YAML
 
 import nemo
 import nemo.collections.asr as nemo_asr
 from nemo.collections.asr.parts import AudioLabelDataset, WaveformFeaturizer, collections, parsers, perturb
 from nemo.core import DeviceType
-from tests.common_setup import NeMoUnitTest
 
 logging = nemo.logging
 
@@ -34,7 +35,8 @@ logging = nemo.logging
 freq = 16000
 
 
-class TestSpeechCommandsPytorch(NeMoUnitTest):
+@pytest.mark.usefixtures("neural_factory")
+class TestSpeechCommandsPytorch(TestCase):
     labels = [
         "cat",
         "dog",
@@ -79,125 +81,7 @@ class TestSpeechCommandsPytorch(NeMoUnitTest):
         if os.path.exists(os.path.join(data_folder, "speech_commands")):
             shutil.rmtree(os.path.join(data_folder, "speech_commands"))
 
-    def test_pytorch_audio_dataset_with_perturbation(self):
-        perturbations = [
-            perturb.WhiteNoisePerturbation(min_level=-90, max_level=-46),
-            perturb.ShiftPerturbation(min_shift_ms=-5.0, max_shift_ms=5.0),
-        ]
-
-        # Execute perturbations with 100% probability
-        prob_perturb = [(1.0, p) for p in perturbations]
-
-        audio_augmentor = perturb.AudioAugmentor(prob_perturb)
-
-        featurizer = WaveformFeaturizer(
-            sample_rate=self.featurizer_config['sample_rate'],
-            int_values=self.featurizer_config['int_values'],
-            augmentor=audio_augmentor,
-        )
-        ds = AudioLabelDataset(manifest_filepath=self.manifest_filepath, labels=self.labels, featurizer=featurizer,)
-
-        for i in range(len(ds)):
-            logging.info(ds[i])
-            # logging.info(ds[i][0].shape)
-            # self.assertEqual(freq, ds[i][0].shape[0])
-
-    def test_dataloader(self):
-        batch_size = 2
-        dl = nemo_asr.AudioToSpeechLabelDataLayer(
-            # featurizer_config=self.featurizer_config,
-            manifest_filepath=self.manifest_filepath,
-            labels=self.labels,
-            batch_size=batch_size,
-            # placement=DeviceType.GPU,
-            sample_rate=16000,
-        )
-        for ind, data in enumerate(dl.data_iterator):
-            # With num_workers update, this is no longer true
-            # Moving to GPU is handled by AudioPreprocessor
-            # data is on GPU
-            # self.assertTrue(data[0].is_cuda)
-            # self.assertTrue(data[1].is_cuda)
-            # self.assertTrue(data[2].is_cuda)
-            # self.assertTrue(data[3].is_cuda)
-            # first dimension is batch
-            self.assertTrue(data[0].size(0) == batch_size)
-            self.assertTrue(data[1].size(0) == batch_size)
-            self.assertTrue(data[2].size(0) == batch_size)
-            self.assertTrue(data[3].size(0) == batch_size)
-
-    def test_trim_silence(self):
-        batch_size = 2
-        normal_dl = nemo_asr.AudioToSpeechLabelDataLayer(
-            # featurizer_config=self.featurizer_config,
-            manifest_filepath=self.manifest_filepath,
-            labels=self.labels,
-            batch_size=batch_size,
-            # placement=DeviceType.GPU,
-            drop_last=False,
-            shuffle=False,
-        )
-        trimmed_dl = nemo_asr.AudioToSpeechLabelDataLayer(
-            # featurizer_config=self.featurizer_config,
-            manifest_filepath=self.manifest_filepath,
-            trim_silence=True,
-            labels=self.labels,
-            batch_size=batch_size,
-            # placement=DeviceType.GPU,
-            drop_last=False,
-            shuffle=False,
-        )
-        for norm, trim in zip(normal_dl.data_iterator, trimmed_dl.data_iterator):
-            for point in range(batch_size):
-                self.assertTrue(norm[1][point].data >= trim[1][point].data)
-
-    def test_audio_preprocessors(self):
-        batch_size = 2
-        dl = nemo_asr.AudioToSpeechLabelDataLayer(
-            # featurizer_config=self.featurizer_config,
-            manifest_filepath=self.manifest_filepath,
-            labels=self.labels,
-            batch_size=batch_size,
-            # placement=DeviceType.GPU,
-            drop_last=False,
-            shuffle=False,
-        )
-
-        installed_torchaudio = True
-        try:
-            import torchaudio
-        except ModuleNotFoundError:
-            installed_torchaudio = False
-            with self.assertRaises(ModuleNotFoundError):
-                to_spectrogram = nemo_asr.AudioToSpectrogramPreprocessor(n_fft=400, window=None)
-            with self.assertRaises(ModuleNotFoundError):
-                to_mfcc = nemo_asr.AudioToMFCCPreprocessor(n_mfcc=15)
-
-        if installed_torchaudio:
-            to_spectrogram = nemo_asr.AudioToSpectrogramPreprocessor(n_fft=400, window=None)
-            to_mfcc = nemo_asr.AudioToMFCCPreprocessor(n_mfcc=15)
-
-        to_melspec = nemo_asr.AudioToMelSpectrogramPreprocessor(features=50)
-
-        for batch in dl.data_iterator:
-            input_signals, seq_lengths, _, _ = batch
-            input_signals = input_signals.to(to_melspec._device)
-            seq_lengths = seq_lengths.to(to_melspec._device)
-
-            melspec = to_melspec.forward(input_signals, seq_lengths)
-
-            if installed_torchaudio:
-                spec = to_spectrogram.forward(input_signals, seq_lengths)
-                mfcc = to_mfcc.forward(input_signals, seq_lengths)
-
-            # Check that number of features is what we expect
-            self.assertTrue(melspec[0].shape[1] == 50)
-
-            if installed_torchaudio:
-                self.assertTrue(spec[0].shape[1] == 201)  # n_fft // 2 + 1 bins
-                self.assertTrue(mfcc[0].shape[1] == 15)
-
-    # @unittest.skip("Init parameters of nemo_asr.AudioToMelSpectrogramPreprocessor are invalid")
+    @pytest.mark.acceptance
     def test_jasper_training(self):
         with open(
             os.path.abspath(os.path.join(os.path.dirname(__file__), "../data/quartznet_speech_recognition.yaml"))
@@ -240,12 +124,12 @@ class TestSpeechCommandsPytorch(NeMoUnitTest):
             tensors=[loss], print_func=lambda x: logging.info(f'Train Loss: {str(x[0].item())}'),
         )
         # Instantiate an optimizer to perform `train` action
-        optimizer = self.nf.get_trainer()
+        optimizer = nemo.backends.pytorch.actions.PtActions()
         optimizer.train(
             [loss], callbacks=[callback], optimizer="sgd", optimization_params={"num_epochs": 10, "lr": 0.0003},
         )
 
-    # @unittest.skip("Init parameters of nemo_asr.AudioToMelSpectrogramPreprocessor are invalid")
+    @pytest.mark.acceptance
     def test_double_jasper_training(self):
         with open(
             os.path.abspath(os.path.join(os.path.dirname(__file__), "../data/quartznet_speech_recognition.yaml"))
@@ -296,11 +180,12 @@ class TestSpeechCommandsPytorch(NeMoUnitTest):
             tensors=[loss], print_func=lambda x: logging.info(str(x[0].item()))
         )
         # Instantiate an optimizer to perform `train` action
-        optimizer = self.nf.get_trainer()
+        optimizer = nemo.backends.pytorch.actions.PtActions()
         optimizer.train(
             [loss], callbacks=[callback], optimizer="sgd", optimization_params={"num_epochs": 10, "lr": 0.0003},
         )
 
+    @pytest.mark.acceptance
     def test_stft_conv(self):
         with open(
             os.path.abspath(os.path.join(os.path.dirname(__file__), "../data/quartznet_speech_recognition.yaml"))
@@ -342,11 +227,12 @@ class TestSpeechCommandsPytorch(NeMoUnitTest):
             tensors=[loss], print_func=lambda x: logging.info(str(x[0].item()))
         )
         # Instantiate an optimizer to perform `train` action
-        optimizer = self.nf.get_trainer()
+        optimizer = nemo.backends.pytorch.actions.PtActions()
         optimizer.train(
             [loss], callbacks=[callback], optimizer="sgd", optimization_params={"num_epochs": 10, "lr": 0.0003},
         )
 
+    @pytest.mark.acceptance
     def test_jasper_eval(self):
         with open(
             os.path.abspath(os.path.join(os.path.dirname(__file__), "../data/quartznet_speech_recognition.yaml"))
