+++ tests/acceptance/test_tts.py
@@ -18,16 +18,19 @@
 
 import os
 import tarfile
+from unittest import TestCase
+
+import pytest
 
 import nemo
 import nemo.collections.asr as nemo_asr
 import nemo.collections.tts as nemo_tts
-from tests.common_setup import NeMoUnitTest
 
 logging = nemo.logging
 
 
-class TestTTSPytorch(NeMoUnitTest):
+@pytest.mark.usefixtures("neural_factory")
+class TestTTSPytorch(TestCase):
     labels = [
         " ",
         "a",
@@ -72,6 +75,7 @@ class TestTTSPytorch(NeMoUnitTest):
         else:
             logging.info("speech data found in: {0}".format(data_folder + "asr"))
 
+    @pytest.mark.acceptance
     def test_tacotron2_training(self):
         data_layer = nemo_asr.AudioToTextDataLayer(
             manifest_filepath=self.manifest_filepath, labels=self.labels, batch_size=4,
@@ -143,6 +147,7 @@ class TestTTSPytorch(NeMoUnitTest):
             [loss_t], callbacks=[callback], optimizer="sgd", optimization_params={"num_epochs": 10, "lr": 0.0003},
         )
 
+    @pytest.mark.acceptance
     def test_waveglow_training(self):
         data_layer = nemo_tts.AudioDataLayer(
             manifest_filepath=self.manifest_filepath, n_segments=4000, batch_size=4, sample_rate=16000
